Move Time object into time_conversion.
